# internal/api/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/a"
			if l := len("/a"); len(elem) >= l && elem[0:l] == "/a" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 's': // Prefix: "ssertion"
				if l := len("ssertion"); len(elem) >= l && elem[0:l] == "ssertion" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleInitializeAssertionRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleFinalizeAssertionRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,POST")
					}

					return
				}
			case 't': // Prefix: "ttestation"
				if l := len("ttestation"); len(elem) >= l && elem[0:l] == "ttestation" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleInitializeAttestationRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleFinalizeAttestationRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,POST")
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/json"
					if l := len("/json"); len(elem) >= l && elem[0:l] == "/json" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleInitializeAttestationJSONRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET")
						}

						return
					}
				}
			}
		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name        string
	summary     string
	operationID string
	pathPattern string
	count       int
	args        [0]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/a"
			if l := len("/a"); len(elem) >= l && elem[0:l] == "/a" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 's': // Prefix: "ssertion"
				if l := len("ssertion"); len(elem) >= l && elem[0:l] == "ssertion" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						// Leaf: InitializeAssertion
						r.name = "InitializeAssertion"
						r.summary = "Initialize Assertion"
						r.operationID = "initializeAssertion"
						r.pathPattern = "/assertion"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						// Leaf: FinalizeAssertion
						r.name = "FinalizeAssertion"
						r.summary = "Finalize Assertion"
						r.operationID = "finalizeAssertion"
						r.pathPattern = "/assertion"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
			case 't': // Prefix: "ttestation"
				if l := len("ttestation"); len(elem) >= l && elem[0:l] == "ttestation" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = "InitializeAttestation"
						r.summary = "Initialize Attestation"
						r.operationID = "initializeAttestation"
						r.pathPattern = "/attestation"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = "FinalizeAttestation"
						r.summary = "Finalize Attestation"
						r.operationID = "finalizeAttestation"
						r.pathPattern = "/attestation"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/json"
					if l := len("/json"); len(elem) >= l && elem[0:l] == "/json" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							// Leaf: InitializeAttestationJSON
							r.name = "InitializeAttestationJSON"
							r.summary = "Initialize Attestation JSON"
							r.operationID = "initializeAttestationJSON"
							r.pathPattern = "/attestation/json"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
				}
			}
		}
	}
	return r, false
}